refactor(is): rename isNullOrUndefined to isNil
