+++ packages/is/src/index.ts
@@ -20,9 +20,9 @@ export * from './lib/isGeneratorFunction';
 export * from './lib/isImmutable';
 export * from './lib/isJSON';
 export * from './lib/isMap';
+export * from './lib/isNil';
 export * from './lib/isNodeContext';
 export * from './lib/isNull';
-export * from './lib/isNullOrUndefined';
 export * from './lib/isNumber';
 export * from './lib/isObject';
 export * from './lib/isPlainObject';
