+++ packages/is/src/lib/isFalsy.spec.ts
@@ -8,6 +8,7 @@ export const falsyValues = [
   'negativeZero',
   'null',
   'undefined',
+  'void0',
   'zero',
   'zeroBigInt',
 ];
