+++ packages/is/src/lib/isImmutable.ts
@@ -1,4 +1,4 @@
-import { isPrimitive } from '..';
+import { isPrimitive } from './isPrimitive';
 
 /**
  * Is the value immutable?
